backend/pyproject.toml
[project]
name = "travel-advisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "pydantic[email]==2.10.6",
    "pydantic-core>=2.27.1",
    "pydantic-settings==2.6.1",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "streamlit>=1.28.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "langchain-community>=0.0.10",
    "openai>=1.3.0",
    "chromadb>=0.4.0",
    "requests>=2.31.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[argon2]==1.7.4",
    "pgvector>=0.4.1",
    "langgraph>=0.6.7",
    "redis<6.0",
    "async-timeout>=5.0.1",
    "pypdf>=4.4.0",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.15",
    "pytest>=8.3.4",
    "ruff>=0.8.5",
]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["migrations"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 50

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
